Need approval for the Code of Conduct changes
The change to requiring Ansible's Code of Conduct wasn't approved:
https://github.com/ansible-collections/overview/pull/132/commits/e574a2563e06a4fa6e75587760564ab8de11988e

However, the checklist in its entirety was approved.

This puts this bit of text in an odd situation where it might be
approved or might not.  This PR is not necessarily the proposal; rather
the change needs to be discussed, possible a new prposal created, and
voted on at the next meeting. Note that both tadeboro and abadger1999
dissent with the change from "Requires a CoC" to "Requires the Ansible
CoC".