Need approval of the CI changes
The substantive changes in
https://github.com/ansible-collections/overview/pull/132/commits/b4c59b2f96338dab54e7bef0f400d678f9fc12c4
were not approved in the meeting.

However, the checklist as a whole was approved in the next meeting.

This leaves the text in this PR in an odd situation where it might be
approved or might not.  This PR is not necessarily the proposal; tather
the change needs to be discussed, possibly a new proposal needs to be
created, and it needs to be voted approved at the next meeting.